requirements.txt
python-dotenv
newsapi
openai